Almost there

+++ src/models/request.rs
@@ -16,7 +16,7 @@ pub struct ChatCompletionRequest {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub function_call: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub temperature: Option<f32>,
+    pub temperature: Option<f64>,
 }
 
 impl Default for ChatCompletionRequest {
